+++ R/symDMatrix.R
@@ -10,15 +10,14 @@
 #'
 #' Each block is a matrix-like block. Currently supported are `ff` objects.
 #' Because the matrix is symmetric, only the diagonal and upper-triangular
-#' blocks are stored.
+#' blocks are stored. The other blocks are inverses of the upper-triangular
+#' blocks (ff views).
 #'
-#' @slot data A nested list in the form of \code{[[G11, G12, G13, ..., G1q],
-#' [G22, G23, ..., G2q], [...], [Gqq]]}, each list element representing a block
-#' of a partitioned symmetric and upper-triangular matrix. There are `(q * (q +
-#' 1)) / 2` blocks in total and each list level consists of `q - i` blocks,
-#' where `q` is the number of column/row blocks and `i` is the current row
-#' index. All blocks except the ones in the last column/row are expected to
-#' have the same dimensions.
+#' @slot blocks A nested list in the form of \code{[[G11, G12, G13, ..., G1q],
+#' [G21, G22, G23, ..., G2q], [...], [G1q, ..., Gqq]]}, each list element
+#' representing a block of a partitioned symmetric and upper-triangular matrix.
+#' All blocks except the ones in the last column/row are expected to have the
+#' same dimensions.
 #' @slot centers A numeric vector storing the values used for column centering
 #' when creating the symmetric matrix.
 #' @slot scales A numeric vector storing the values used for column scaling
@@ -30,7 +29,7 @@
 #' @aliases symDMatrix-class
 #' @export symDMatrix
 #' @exportClass symDMatrix
-symDMatrix <- setClass("symDMatrix", slots = c(data = "list", centers = "numeric", scales = "numeric"))
+symDMatrix <- setClass("symDMatrix", slots = c(centers = "numeric", scales = "numeric"), contains = "RowLinkedMatrix")
 
 
 #' Create a New symDMatrix Instance.
@@ -49,7 +48,6 @@ symDMatrix <- setClass("symDMatrix", slots = c(data = "list", centers = "numeric
 #' @param .Object The [symDMatrix-class] instance to be initialized. This
 #' argument is passed in by R and can be ignored, but still needs to be
 #' documented.
-#' @param data A nested list of blocks.
 #' @param centers A numeric vector storing the values used for column centering
 #' when creating the symmetric matrix.
 #' @param scales A numeric vector storing the values used for column scaling
@@ -59,215 +57,55 @@ symDMatrix <- setClass("symDMatrix", slots = c(data = "list", centers = "numeric
 #' @seealso [as.symDMatrix()] to create a [symDMatrix-class] object from other
 #' objects.
 #' @export
-setMethod("initialize", "symDMatrix", function(.Object, data, centers = 0L, scales = 1L) {
-    nBlocks <- length(data)
-    # Test that there is at least one block
+setMethod("initialize", "symDMatrix", function(.Object, ..., centers = 0L, scales = 1L) {
+    blocks <- list(...)
+    nBlocks <- length(blocks)
+    # Stop if there are no blocks
     if (nBlocks == 0L) {
-        stop("data needs to contain at least one block")
+        stop("there needs to be at least one block")
     }
-    # Test that data has the right structure
-    blocksPerRow <- sapply(data, length)
-    if (!identical(blocksPerRow, seq(nBlocks, 1L))) {
-        stop("data needs to be a nested list in the following structure: [[G11, G12, G13, ..., G1q], [G22, G23, ..., G2q], [...], [Gqq]]")
+    # Stop if blocks are not of type ColumnLinkedMatrix
+    if (!all(sapply(blocks, class) == "ColumnLinkedMatrix")) {
+        stop("blocks need to be of type ColumnLinkedMatrix")
     }
-    # Block-level tests
-    rowDims <- rep(NA_integer_, nBlocks)
+    # Stop if the number of nested blocks is inconsistent
+    if (length(unique(sapply(blocks, LinkedMatrix::nNodes))) > 1L) {
+        stop("number of nested blocks is inconsistent")
+    }
+    rowDims <- sapply(blocks, nrow)
     colDims <- rep(NA_integer_, nBlocks)
-    for (i in seq(0L, nBlocks - 1L)) {
-        for (j in seq(1L, nBlocks - i)) {
-            block <- data[[i + 1L]][[j]]
-            # Test that all blocks are matrix-like
-            if (!isMatrixLike(block)) {
-                stop("data: all blocks need to be matrix-like objects")
-            }
-            # Test that all blocks per row have the same number of rows
-            if (is.na(rowDims[i + 1L])) {
-                rowDims[i + 1L] <- nrow(block)
-            } else {
-                if (nrow(block) != rowDims[i + 1L]) {
-                    stop("data: all blocks per row need the same number of rows")
-                }
-            }
-            # Test that all blocks per column have the same number of columns
-            if (is.na(colDims[j + i])) {
-                colDims[j + i] <- ncol(block)
-            } else {
-                if (ncol(block) != colDims[j + i]) {
-                    stop("data: all blocks per column need the same number of columns")
-                }
+    for (rowIndex in seq_len(nBlocks)) {
+        rowBlocks <- blocks[[rowIndex]]
+        # Stop if nested blocks do not inherit from ff_matrix
+        if (!all(sapply(rowBlocks, inherits, "ff_matrix"))) {
+            stop("nested blocks need to inherit from ff_matrix")
+        }
+        # Stop if nested blocks do not have the same number of columns per column of blocks
+        if (all(is.na(colDims))) {
+            colDims[] <- sapply(rowBlocks, ncol)
+        } else {
+            if (!all(sapply(rowBlocks, ncol) == colDims)) {
+                stop("all nested blocks need the same number of columns per column of blocks")
             }
         }
     }
     if (nBlocks > 1L) {
-        # Test that non-final blocks are square
+        # Stop if non-final block is not square
         if (any(rowDims[-length(rowDims)] != colDims[-length(colDims)])) {
-            stop("data: non-final blocks need to be square")
+            stop("non-final blocks need to be square")
         }
     } else {
-        # Test that the first block is square
-        if (nrow(data[[1]][[1]]) != ncol(data[[1]][[1]])) {
-            stop("data: the first block needs to be square")
+        # Stop if first block is not square
+        if (nrow(blocks[[1]][[1]]) != ncol(blocks[[1]][[1]])) {
+            stop("the first block needs to be square")
         }
     }
-    .Object@data <- data
+    # Call RowLinkedMatrix constructor
+    .Object <- callNextMethod(.Object, ...)
     .Object@centers <- centers
     .Object@scales <- scales
     return(.Object)
 })
-
-
-#' @export
-is.matrix.symDMatrix <- function(x) {
-    TRUE
-}
-
-
-#' @export
-dim.symDMatrix <- function(x) {
-    p <- sum(sapply(x@data[[1L]], ncol))
-    c(p, p)
-}
-
-
-#' @export
-length.symDMatrix <- function(x) {
-    prod(dim(x))
-}
-
-
-names.symDMatrix <- function(x) {
-    blockNames <- lapply(x@data[[1L]], function(block) {
-        colnames(block)
-    })
-    isNULL <- sapply(blockNames, function(blockName) {
-        is.null(blockName)
-    })
-    if (any(isNULL)) {
-        NULL
-    } else {
-        unlist(blockNames)
-    }
-}
-
-
-#' @export
-dimnames.symDMatrix <- function(x) {
-    names <- names.symDMatrix(x)
-    if (is.null(names)) {
-        NULL
-    } else {
-        list(names, names)
-    }
-}
-
-
-#' @export
-`[.symDMatrix` <- function(x, i, j, drop = TRUE) {
-
-    nargs <- nargs()
-
-    nX <- nrow(x)
-
-    # Single Index: x[i]
-    if (nargs == 2L && !missing(i) && missing(j)) {
-
-        singleIndex <- TRUE
-
-        n <- 1L
-        p <- length(i)
-
-        # Convert single index to multi index
-        k <- i - 1L
-        paired_i <- k %% nX
-        paired_j <- as.integer(k / nX)
-        paired_i <- paired_i + 1L
-        paired_j <- paired_j + 1L
-
-    # No index and multi Index: x[], or x[, ], and x[i, j], x[i, ], or x[, j]
-    } else {
-
-        singleIndex <- FALSE
-
-        pX <- ncol(x)
-
-        if (missing(i)) {
-            i <- 1L:nX
-        } else if (typeof(i) == "logical") {
-            i <- rep_len(i, nX)
-            i <- which(i)
-        } else if (typeof(i) == "character") {
-            i <- match(i, rownames(x))
-        } else if (typeof(i) == "double") {
-            i <- as.integer(i)
-        }
-        if (missing(j)) {
-            j <- 1L:pX
-        } else if (typeof(j) == "logical") {
-            j <- rep_len(j, pX)
-            j <- which(j)
-        } else if (typeof(j) == "character") {
-            j <- match(j, colnames(x))
-        } else if (typeof(j) == "double") {
-            j <- as.integer(j)
-        }
-
-        n <- length(i)
-        p <- length(j)
-
-        # Create all combinations of i and j
-        paired_i <- rep(i, each = p)
-        paired_j <- rep(j, times = n)
-
-    }
-
-    # Switch indices for combinations in which i is larger than j to redirect
-    # queries to the lower triangle to the upper triangle
-    switch <- paired_i > paired_j
-    flip <- paired_i[switch]
-    paired_i[switch] <- paired_j[switch]
-    paired_j[switch] <- flip
-
-    # Retrieve block size
-    blockSize <- blockSize(x)
-
-    # Create retrieval index
-    row_blocks <- as.integer(ceiling(paired_i / blockSize))
-    col_blocks <- as.integer(ceiling(paired_j / blockSize))
-    local_i <- paired_i - (row_blocks - 1L) * blockSize
-    local_j <- paired_j - (col_blocks - 1L) * blockSize
-
-    # Initialize output matrix
-    names <- names.symDMatrix(x)
-    if (!is.null(names) && !singleIndex) {
-        dimnames <- list(names[i], names[j])
-    } else {
-        dimnames <- NULL
-    }
-    OUT <- matrix(data = double(), nrow = n, ncol = p, dimnames = dimnames)
-
-    # Create output index
-    out_i <- rep(1L:n, each = p)
-    out_j <- rep(1L:p, times = n)
-
-    # Retrieve elements by block
-    for (row_block in unique(row_blocks)) {
-        row_block_matches <- row_blocks == row_block
-        for (col_block in unique(col_blocks[row_block_matches])) {
-            cur_block <- row_block_matches & col_blocks == col_block
-            block <- x@data[[row_block]][[col_block - row_block + 1L]]
-            out_idx <- (out_j[cur_block] - 1L) * n + out_i[cur_block]
-            local_idx <- (local_j[cur_block] - 1L) * nrow(block) + local_i[cur_block]
-            OUT[out_idx] <- block[local_idx]
-        }
-    }
-
-    if (drop == TRUE && (n == 1L || p == 1L)) {
-        return(OUT[, ])
-    } else {
-        return(OUT)
-    }
-
-}
 
 
 #' Load symDMatrix Objects from .RData Files.
@@ -290,8 +128,8 @@ load.symDMatrix <- function(file, envir = parent.frame()) {
         if (class(object) == "symDMatrix") {
             nBlocks <- nBlocks(object)
             for (i in 1L:nBlocks) {
-                for (j in i:nBlocks) {
-                    object@data[[i]][[j - i + 1L]] <- initializeBlock(object@data[[i]][[j - i + 1L]], path = dirname(file))
+                for (j in 1L:nBlocks) {
+                    object[[i]][[j]] <- initializeBlock(object[[i]][[j]], path = dirname(file))
                 }
             }
         }
@@ -310,9 +148,6 @@ initializeBlock <- function(x, ...) {
 # Absolute paths to ff files are not stored, so the ff objects have to be
 # loaded from the same directory as the RData file.
 initializeBlock.ff_matrix <- function(x, path, ...) {
-    if (!requireNamespace("ff", quietly = TRUE)) {
-        stop("The ff package is needed for this function to work. Please install it.", call. = FALSE)
-    }
     # Store current working directory and set working directory to path
     cwd <- getwd()
     setwd(path)
@@ -338,7 +173,7 @@ initializeBlock.default <- function(x, ...) {
 #' @example man/examples/nBlocks.R
 #' @export
 nBlocks <- function(x) {
-    length(x@data[[1L]])
+    LinkedMatrix::nNodes(x)
 }
 
 
@@ -357,9 +192,9 @@ nBlocks <- function(x) {
 #' @example man/examples/blockSize.R
 #' @export
 blockSize <- function(x, last = FALSE) {
-    row <- x@data[[1L]]
+    row <- x[[1L]]
     if (last) {
-        ncol(row[[length(row)]])
+        ncol(row[[LinkedMatrix::nNodes(row)]])
     } else {
         ncol(row[[1L]])
     }
@@ -376,13 +211,13 @@ blockSize <- function(x, last = FALSE) {
 #' the end index.
 #' @export
 blockIndex <- function(x) {
-    n <- length(x@data)
-    index <- matrix(nrow = n, ncol = 3L)
+    nNodes <- LinkedMatrix::nNodes(x)
+    index <- matrix(nrow = nNodes, ncol = 3L)
     colnames(index) <- c("block", "ini", "end")
     end <- 0L
-    for (i in 1L:n) {
+    for (i in 1L:nNodes) {
         ini <- end + 1L
-        end <- ini + nrow(x@data[[i]][[1L]]) - 1L
+        end <- ini + nrow(x[[i]][[1L]]) - 1L
         index[i, ] <- c(i, ini, end)
     }
     return(index)
@@ -426,59 +261,49 @@ as.symDMatrix <- function(x, ...) {
 #' @seealso [load.symDMatrix()] to reload the [symDMatrix-class] object.
 #' @export
 as.symDMatrix.matrix <- function(x, blockSize = 5000L, vmode = "double", folderOut = randomString(), ...) {
-
-    if (!requireNamespace("ff", quietly = TRUE)) {
-        stop("The ff package is needed for this function to work. Please install it.", call. = FALSE)
-    }
-
     n <- nrow(x)
-
     if (ncol(x) != n) {
         stop("x must be a square matrix")
     }
-
     if (file.exists(folderOut)) {
         stop(folderOut, " already exists")
     }
     dir.create(folderOut)
-
     # Determine number of blocks from block size
     nBlocks <- as.integer(ceiling(nrow(x) / blockSize))
-
     # Determine subjects of each block
-    index <- matrix(data = integer(), nrow = nBlocks, ncol = 3L)
-    index[1L, ] <- c(1L, 1L, min(n, blockSize))
+    index <- matrix(data = integer(), nrow = nBlocks, ncol = 2L)
+    index[1L, ] <- c(1L, min(n, blockSize))
     if (nBlocks > 1L) {
         for (i in 2L:nBlocks) {
-            index[i, 1L] <- i
-            index[i, 2L] <- index[(i - 1L), 3L] + 1L
-            index[i, 3L] <- min(index[i, 2L] + blockSize - 1L, n)
+            index[i, 1L] <- index[(i - 1L), 2L] + 1L
+            index[i, 2L] <- min(index[i, 1L] + blockSize - 1L, n)
         }
     }
-
     # Create nested list
-    dataList <- vector(mode = "list", length = nBlocks)
-    for (r in 1L:nBlocks) {
-        rowIndex <- seq(index[r, 2L], index[r, 3L])
-        dataList[[r]] <- vector(mode = "list", length = nBlocks - r)
-        for (s in r:nBlocks) {
-            colIndex <- seq(index[s, 2L], index[s, 3L])
-            blockName <- paste0("data_", padDigits(r, nBlocks), "_", padDigits(s, nBlocks), ".bin")
-            block <- ff::ff(dim = c(length(rowIndex), length(colIndex)), vmode = vmode, initdata = x[rowIndex, colIndex], filename = paste0(folderOut, "/", blockName))
-            colnames(block) <- colnames(x)[colIndex]
-            rownames(block) <- rownames(x)[rowIndex]
+    args <- vector(mode = "list", length = nBlocks)
+    for (rowIndex in 1L:nBlocks) {
+        rowRanges <- seq(index[rowIndex, 1L], index[rowIndex, 2L])
+        rowArgs <- vector(mode = "list", length = nBlocks)
+        for (colIndex in 1L:nBlocks) {
+            colRanges <- seq(index[colIndex, 1L], index[colIndex, 2L])
+            blockName <- paste0("data_", padDigits(rowIndex, nBlocks), "_", padDigits(colIndex, nBlocks), ".bin")
+            block <- ff::ff(dim = c(length(rowRanges), length(colRanges)), vmode = vmode, initdata = x[rowRanges, colRanges], filename = paste0(folderOut, "/", blockName), dimnames = list(rownames(x)[rowRanges], colnames(x)[colRanges]))
             # Change ff path to a relative one
             bit::physical(block)$filename <- blockName
-            dataList[[r]][[s - r + 1L]] <- block
+            if (colIndex >= rowIndex) {
+                rowArgs[[colIndex]] <- block
+            } else {
+                rowArgs[[colIndex]] <- ff::vt(args[[colIndex]][[rowIndex]])
+            }
         }
+        args[[rowIndex]] <- do.call(LinkedMatrix::ColumnLinkedMatrix, rowArgs)
     }
-
-    # Create symDMatrix object from nested list
-    symDMatrix <- symDMatrix(data = dataList, centers = 0L, scales = 1L)
-
-    # Save RData object
+    # Append centers and scales to args
+    args <- c(args, list(centers = 0L, scales = 1L))
+    # Create symDMatrix object from args
+    symDMatrix <- do.call(symDMatrix, args)
     save(symDMatrix, file = "symDMatrix.RData")
-
     return(symDMatrix)
 }
 
@@ -506,32 +331,40 @@ as.symDMatrix.matrix <- function(x, blockSize = 5000L, vmode = "double", folderO
 #' @export
 as.symDMatrix.character <- function(x, centers = 0L, scales = 1L, ...) {
     nBlocks <- as.integer((-1L + sqrt(1L + 4L * 2L * length(x))) / 2L)
-    dataList <- vector(mode = "list", length = nBlocks)
+    args <- vector(mode = "list", length = nBlocks)
     counter <- 1L
     for (i in 1L:nBlocks) {
-        dataList[[i]] <- vector(mode = "list", length = nBlocks - i)
-        for (j in i:nBlocks) {
-            loadingEnv <- new.env()
-            file <- x[[counter]]
-            load(file = file, envir = loadingEnv)
-            names <- ls(envir = loadingEnv)
-            # Make sure that at least and at most one object is matrix-like
-            isMatrixLike <- sapply(names, function(name) {
-                object <- get(name, envir = loadingEnv)
-                isMatrixLike(object)
-            })
-            if (sum(isMatrixLike) != 1L) {
-                stop("only one object per RData file can be a matrix-like object")
+        rowArgs <- vector(mode = "list", length = nBlocks)
+        for (j in 1L:nBlocks) {
+            if (j >= i) {
+                loadingEnv <- new.env()
+                file <- x[[counter]]
+                load(file = file, envir = loadingEnv)
+                names <- ls(envir = loadingEnv)
+                # Make sure that exactly one object inherits from ff
+                inherits <- sapply(names, function(name) {
+                    object <- get(name, envir = loadingEnv)
+                    inherits(object, "ff_matrix")
+                })
+                if (sum(inherits) != 1L) {
+                    stop("only one object per RData file can inherit from ff_matrix")
+                }
+                object <- get(names[which(inherits)], envir = loadingEnv)
+                # Initialize the matrix-like object
+                object <- initializeBlock(object, path = dirname(file))
+                rowArgs[[j]] <- object
+                counter <- counter + 1L
+            } else {
+                rowArgs[[j]] <- ff::vt(args[[j]][[i]])
             }
-            object <- get(names[which(isMatrixLike)], envir = loadingEnv)
-            # Initialize the matrix-like object
-            object <- initializeBlock(object, path = dirname(file))
-            dataList[[i]][[j - i + 1L]] <- object
-            counter <- counter + 1L
         }
+        args[[i]] <- do.call(LinkedMatrix::ColumnLinkedMatrix, rowArgs)
     }
-    G <- symDMatrix(data = dataList, centers = centers, scales = scales)
-    return(G)
+    # Append centers and scales to args
+    args <- c(args, list(centers = 0L, scales = 1L))
+    # Create symDMatrix object from args
+    symDMatrix <- do.call(symDMatrix, args)
+    return(symDMatrix)
 }
 
 
@@ -542,9 +375,4 @@ randomString <- function(n = 10L) {
 
 padDigits <- function(x, total) {
     formatC(x, width = as.integer(log10(total) + 1L), format = "d", flag = "0")
-}
-
-
-isMatrixLike <- function(x) {
-    length(dim(x)) == 2L
 }